% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_recursives.R
\name{followers_list_recursive}
\alias{followers_list_recursive}
\title{Recursively Get Followers List}
\usage{
followers_list_recursive(screen_name, user_id, data.connection,
  authentication.vector, skip_status = NULL,
  include_user_entities = NULL, tweet_mode = "extended", ...)
}
\arguments{
\item{screen_name}{character single screen_name.  This is only used is \code{user_id} is missing.}

\item{user_id}{numeric or character single user_id.}

\item{data.connection}{a DBI connection to an RSQLite \code{\link{twitter_database}}.}

\item{authentication.vector}{character vector containing authentication tokens and secrets.}

\item{skip_status}{logical indicating whether to exclude the user status from the returned
follower user objects.}

\item{include_user_entities}{logical indicating whether to include profile entities (e.g. profile url)
in the returned follower user objects.}

\item{tweet_mode}{character either 'extended' for full_text statuses or 'compat' for
140 character compatability.}

\item{...}{other named parameters passed to \code{\link{insert_users}}.}
}
\value{
character collected \code{user_id} if \code{data.connection} is supplied, otherwise a list of user objects.
}
\description{
Recursively call the Followers List API to get all followers
}
\details{
This function calls recursively calls the
followers/list endpoint (see
\href{https://developer.twitter.com/en/docs/accounts-and-users/follow-search-get-users/api-reference/get-followers-list}{Followers List API Documentation}).
A user_id or screen_name is required.  If an authentication vector is not
provided and there is no globally defined \code{auth.vector}, an error is returned.  A *one minute* delay
is included for each query in order to avoid exceeding the rate limit.
}
\examples{

\dontrun{
auth.vector <- authorize_IT()

dma.relief.followers <- followers_list_recursive(
  "dominica_relief",
  authentication.vector = auth.vector
)

cat(
  paste(
    "@",
    sapply(dma.relief.followers[1:20],function(x) return(x$screen_name)),
    ": ",
    sapply(dma.relief.followers[1:20],function(x) return(x$description)),
    sep="",
    collapse="\\n"
  )
)
}
}
\seealso{
\code{\link{followers_list}}, \code{\link{insert_followers}}, \code{\link{insert_users}}, \code{\link{twitter_database}}
}
