% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{friends_list}
\alias{friends_list}
\title{Get Friends List}
\usage{
friends_list(screen_name, user_id, authentication.vector, cursor = NULL,
  count = 200, skip_status = NULL, include_user_entities = NULL,
  tweet_mode = "extended")
}
\arguments{
\item{screen_name}{character single screen_name.  This is only used is \code{user_id} is missing.}

\item{user_id}{numeric or character single user_id.}

\item{authentication.vector}{character vector containing authentication tokens and secrets.
See \code{\link{authorize_app}} and \code{\link{authorize_IT}}.}

\item{cursor}{numeric or character cursor value for controlling pagination.}

\item{count}{numeric number of user_ids per page, up to 200.}

\item{skip_status}{logical indicating whether to exclude the user status from the returned
follower user objects.}

\item{include_user_entities}{logical indicating whether to include profile entities (e.g. profile url)
in the returned follower user objects.}

\item{tweet_mode}{character either 'extended' for full_text statuses or 'compat' for
140 character compatability.}
}
\value{
On success, an R list representation of the returned JSON object.  On failure, a warning
is thrown and the http response is returned, if any.
}
\description{
Get a User's Friends' User Objects Using the Friends List API
}
\details{
This function calls \code{\link{twitter_request}} using the Twitter
friends/list endpoint (see
\href{https://developer.twitter.com/en/docs/accounts-and-users/follow-search-get-users/api-reference/get-friends-list}{Friends List API Documentation}).
A user_id or screen_name is required.  If an authentication vector is not
provided and there is no globally defined \code{auth.vector}, an error is returned.
}
\examples{

\dontrun{
auth.vector <- authorize_IT()

potus.friends <- friends_list(
  "realDonaldTrump",
  authentication.vector = auth.vector
)

cat(
  paste(
    "@",
    sapply(potus.friends$users[1:20],function(x) return(x$screen_name)),
    ": ",
    sapply(potus.friends$users[1:20],function(x) return(x$description)),
    sep="",
    collapse="\\n"
  )
)
}
}
\seealso{
\code{\link{twitter_request}}, \code{\link{authorize_app}}, \code{\link{friends_ids}},
\code{\link{followers_list}}
}
