% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data.R
\name{get_all_followers}
\alias{get_all_followers}
\title{Get users' followers}
\usage{
get_all_followers(con, authentication.vector, user_id = NULL,
  screen_name = NULL, last.query.time.id = NULL,
  last.query.time.list = NULL)
}
\arguments{
\item{con}{Twitter database connection (see \code{\link{twitter_database}}).}

\item{authentication.vector}{character vector with Twitter authentication
keys.  See \code{\link{authorize_IT}} or \code{\link{authorize_app}}.}

\item{user_id}{character vector of Twitter user_ids.}

\item{screen_name}{character vector of Twitter screen_names.  Ignored
if \code{user_id} is supplied.}

\item{last.query.time.id}{POSIXct time of last followers IDs API query.  See
\code{\link{followers_ids_recursive}}.}

\item{last.query.time.list}{POSIXct time of last friends IDs API query.  See
\code{\link{followers_list_recursive}}.}
}
\value{
\code{NULL} (Invisible).
}
\description{
Get all users' followers and insert the relationships into a database.
}
\details{
This function collects all of the followers for each user in the
\code{query_user} table, if they have not already beeen collected, 
 and inserts the relationships into the \code{followers} table.  
If a \code{user_id} or \code{screen_name} vector is supplied, only 
these users' followers are collected.  Additionally, if the supplied
users are not present in the \code{query_users} table, they are inserted.
The Twitter user object for each friend is also collected and inserted
into the \code{user} table. 
If \code{authentication.vector} is not supplied and no global 
\code{auth.vector} exists, this method returns an error.
}
\section{Caution}{

Each API friend query takes one minute, due to rate limiting.  For a 
large list of users or even a small list of users with many followers,
this function can take a very long time to execute.  The number of minutes it
takes to get followers for a single user is 
approximately \code{ceil(followers_count/5000)}.
}

\examples{

\dontrun{
auth.vector <- authorize_IT()

con <- twitter_database(
  "test.sqlite",
  query.users.df = data.frame(
    screen_name = c(
      "pb2pv",
      "zlisto"
    )
  )
)
get_all_followers(con)

DBI::dbDisonnect(con)
}
}
\seealso{
\code{\link{twitter_database}}, \code{\link{followers_ids_recursive}},
\code{\link{insert_friends}}
}
