% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{open_status}
\alias{open_status}
\title{Open a Tweet}
\usage{
open_status(status_id)
}
\arguments{
\item{status_id}{character Twitter status ID.}
}
\value{
NULL (Invisible).
}
\description{
View a Twitter status in a browser
}
\details{
This function opens a browser window to the Tweet corresponding to the \code{status_id} provided.
Authentication tokens are not required.
}
\examples{


open_status("1009830068624019457")

}
\seealso{
\code{\link{open_user}}
}
