% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sentiment_plots}
\alias{sentiment_plots}
\title{Sentiment Plots of Tweets}
\usage{
sentiment_plots(con, start.date = NULL, end.date = NULL,
  where.criteria = NULL, file.name.prefix = NULL, caption = "")
}
\arguments{
\item{con}{SQLite data connection to Twitter database (see \code{\link{twitter_database}}).}

\item{start.date}{character date (or datetime) of earliest statuses queried.}

\item{end.date}{character date.  Results will include statuses 
with creation times or dates strictly less than this parameter.}

\item{where.criteria}{character additional criteria to filter the status results, in 
SQLite syntax.}

\item{file.name.prefix}{character string prepended to file names of plots.}

\item{caption}{character caption to include on the plot.}
}
\value{
NULL (Invisible).
}
\description{
Plot NRC sentiment of Tweets over time.
}
\details{
This function uses the NRC sentiment values for each status assigned
by the \code{syuzhet} package.  This sentiment classification must be
present in the Twitter data connection for this function to be useful.  
The function produces two plots: (1) a bar plot showing absolute sentiment
quantities over time, and (2) a line plot showing relative sentiment values 
over time.  Sentiments are categorized into a Likert-like scale.
}
\examples{
\dontrun{
auth.vector <- authorize_IT()

con <- twitter_database(
  "test.sqlite",
  query.users.df = data.frame(
    screen_name = c(
      "pb2pv",
      "zlisto"
    )
  ),
  query.text.df = data.frame(
    query_text = c(
      "#throwbackthursday",
      "#worstfirstdate"
    )
  )
)

update_user_timelines(con)
update_search(con)

sentiment_plots(
  con,
  start.date = as.Date(Sys.time())-7,
  where.criteria = "user.screen_name = 'zlisto'",
  caption = "@zlisto"
)


DBI::dbDisonnect(con)
}
}
\seealso{
\code{\link{twitter_database}}, \code{\link{authorize_app}},
\code{\link{insert_statuses}}
}
