% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{status_lookup}
\alias{status_lookup}
\title{Lookup Statuses}
\usage{
status_lookup(status_id, authentication.vector, include_entities = TRUE,
  trim_user = FALSE, map = NULL, include_ext_alt_text = NULL,
  include_card_uri = NULL, tweet_mode = "extended")
}
\arguments{
\item{status_id}{numeric or character vector of up to 100 status_ids.}

\item{authentication.vector}{character vector containing authentication tokens and secrets.
See \code{\link{authorize_app}} and \code{\link{authorize_IT}}.}

\item{include_entities}{logical indicating whether to include profile entities (e.g., urls)
as separate nodes in the returned json object.}

\item{trim_user}{logical indicating whether to remove the user object from each status.}

\item{map}{logical indicating whether to return status_ids with NULL values if they do not
exist or cannot be viewed by the authenticated user.}

\item{include_ext_alt_text}{logical indicating whether to return "alt" text assigned to any
attached media entities.}

\item{include_card_uri}{logical indicating whether to include a \code{card_uni} attribute with each
tweet when there is an ad attached to the tweet using the \code{card_uni} value.}

\item{tweet_mode}{character either 'extended' for full_text statuses or 'compat' for
140 character compatability.}
}
\value{
On success, an R list representation of the returned JSON object.  On failure, a warning
is thrown and the http response is returned, if any.
}
\description{
Get Twitter Status Objects from the Status Lookup API
}
\details{
This function calls \code{\link{twitter_request}} using the Twitter
statuses/lookup endpoint (see
\href{https://developer.twitter.com/en/docs/tweets/post-and-engage/api-reference/get-statuses-lookup}{Status Lookup API Documentation}).
A status_id vector is required.  If an authentication vector is not
provided and there is no globally defined \code{auth.vector}, an error is returned.
}
\examples{

\dontrun{
auth.vector <- authorize_IT()

status.list <- status_lookup(
  c(
   "1059843699298000897",
   "1059843147906322432",
   "1059842917769068544",
   "1059842858797223937"
  ),
  authentication.vector = auth.vector
)

cat(status.list[[1]]$full_text)
}
}
\seealso{
\code{\link{twitter_request}}, \code{\link{authorize_app}}, \code{\link{status_show}}
}
