% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{status_media}
\alias{status_media}
\title{Access Tweet Media}
\usage{
status_media(media.table.row, conn, save.to.file = FALSE,
  file.name = NULL, return.image = FALSE, update.media.b64 = TRUE,
  hash.size = 16, display.image = FALSE, showWarnings = TRUE)
}
\arguments{
\item{media.table.row}{data.frame row of the \code{media} table in the Twitter data connection.
The row must contain the \code{media_url} and \code{img_b64} columns.}

\item{conn}{DBI twitter database object.}

\item{save.to.file}{logical indicating whether to save the image to a file.}

\item{file.name}{character name of file to save.  Ignored if \code{save.to.file} is \code{FALSE}.}

\item{return.image}{logical indicating whether to return the \code{image} object to the user.}

\item{update.media.b64}{logical.  If \code{TRUE}, the 64-bit encoded image and average hash value
will be updated in the \code{media} table in the Twitter database connection.}

\item{hash.size}{numeric hash table dimension for average hash algorithm.  See \code{media} table
description in \code{\link{twitter_database}}.}

\item{display.image}{logical.  If \code{TRUE} the image will be rendered on the screen.}

\item{showWarnings}{logical.  If \code{FALSE}, warnings will be suppressed.}
}
\value{
NULL (Invisible).
}
\description{
Function to access and store images attached to Tweets
}
\details{
This function provides capability to download Tweet media or, if a 64-bit encoding
is available in the data connection, to convert it back to an image.  The image
can be saved to a file, encoded and written into the database, and/or rendered on
the screen.
}
\examples{
\dontrun{
auth.vector <- authorize_IT()

con <- twitter_database("throwback.sqlite")
tbt <- search_tweets_recursive(
  "throwbackthursday",
  data.connection = con
)

query <- "SELECT *,COUNT(media_id) AS n FROM media GROUP_BY media_id ORDER BY n DESC LIMIT 1;"
media.df <- DBI::dbGetQuery(con,query)
status_media(
  media.df[1,],
  con,
  save.to.file = FALSE,
  file.name = NULL,
  return.image = NULL,
  update.media.b64 = TRUE,
  hash.size = 16
)
}
}
\seealso{
\code{\link{twitter_database}}, \code{\link{reconstitute_image}}
}
