% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data.R
\name{summarize_database}
\alias{summarize_database}
\title{Get Database Summary Information#}
\usage{
summarize_database(con)
}
\arguments{
\item{con}{Twitter database connection (see \code{\link{twitter_database}}).}
}
\value{
\code{NULL} (Invisible).
}
\description{
Summary information for a Twitter database.
}
\details{
Outputs the numbers of records in each table to the screen, as well as the 
database size.
}
\examples{

\dontrun{
auth.vector <- authorize_IT()

con <- twitter_database(
  "test.sqlite",
  query.users.df = data.frame(
    screen_name = c(
      "pb2pv",
      "zlisto"
    )
  ),
  query.text.df = data.frame(
    query_text = c(
      "#throwbackthursday",
      "#worstfirstdate"
    )
  )
)

get_all_friends(con)
update_user_timelines(con)
update_search(con)

summarize_database(con)

DBI::dbDisonnect(con)
}
}
\seealso{
\code{\link{twitter_database}}
}
