% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{timeplot}
\alias{timeplot}
\title{Cummulative Tweet Time Plot}
\usage{
timeplot(con, start.date = NULL, end.date = NULL,
  where.criteria = NULL, group.column = NULL, file.name = NULL,
  caption = "", log.scale = FALSE)
}
\arguments{
\item{con}{SQLite data connection to Twitter database (see \code{\link{twitter_database}}).}

\item{start.date}{character date (or datetime) of earliest statuses queried.}

\item{end.date}{character date.  Results will include statuses 
with creation times or dates strictly less than this parameter.}

\item{where.criteria}{character additional criteria to filter the status results, in 
SQLite syntax.}

\item{group.column}{character column from Twitter database in the form of \code{table.column}.
Results will be grouped for comparison on this variable.}

\item{file.name}{character name of file to save the word cloud.  If \code{NULL}, the
time plot is displayed on the screen.}

\item{caption}{character caption to include on the plot.}

\item{log.scale}{logical indicating whether to plot the Tweet counts on a log scale.}
}
\value{
NULL (Invisible).
}
\description{
Plot the cummulative number of tweets over time.
}
\details{
Use the \code{group.column} for comparing the Tweet rates of different
groups.
}
\examples{
\dontrun{
auth.vector <- authorize_IT()

con <- twitter_database(
  "test.sqlite",
  query.users.df = data.frame(
    screen_name = c(
      "pb2pv",
      "zlisto"
    )
  ),
  query.text.df = data.frame(
    query_text = c(
      "#throwbackthursday",
      "#worstfirstdate"
    )
  )
)

update_user_timelines(con)
update_search(con)

## All tweets together (not very informative in this example)
timeplot(
  con,
)


timeplot(
  con,
  start.date = Sys.Date() - 7,
  where.criteria = "query_text.id = 1 OR query_text.id = 2",
  group.column = "query_text.queryt_text",
  log.scale = TRUE
)

DBI::dbDisonnect(con)
}
}
\seealso{
\code{\link{twitter_database}}, \code{\link{authorize_app}},
\code{\link{insert_statuses}}
}
