% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinfoR.R
\docType{package}
\name{twinfoR}
\alias{twinfoR}
\alias{twinfoR-package}
\title{twinfoR: A package for mining Twitter.}
\description{
Methods for authorizing Apps and obtaining access tokens,
collecting and storing Twitter data, and analyzing
the data through plots, wordclouds, and descriptive statistics.
Data is collected through the
\href{https://developer.twitter.com/en/docs/api-reference-index}{Twitter REST APIs}.
}
\section{Authentication}{

The \code{\link{authorize_app}} and \code{\link{authorize_IT}} provide
methods to obtain access tokens needed for most API methods.
}

\section{Data}{

}

\section{Collection}{

}

\section{Analysis}{

}

\examples{





}
