% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data.R
\name{update_search}
\alias{update_search}
\title{Search and Insert Tweets}
\usage{
update_search(con, authentication.vector, query.text = NULL, ...)
}
\arguments{
\item{con}{Twitter database connection (see \code{\link{twitter_database}}).}

\item{authentication.vector}{character vector with Twitter authentication
keys.  See \code{\link{authorize_IT}} or \code{\link{authorize_app}}.}

\item{query.text}{character vector of Twitter search query strings.}

\item{...}{additional named parameters passed to \code{\link{search_tweets_recursive}}.}
}
\value{
\code{NULL} (Invisible).
}
\description{
Search and insert new statuses in a Twitter database
}
\details{
This function searches the Twitter API for status objects using 
the keyword search strings in the \code{query_text} table.  Only statuses that
are newer than the most recent status, annotated in the \code{since_id} column,
are searched.  If a \code{query.text} vector
is supplied, then *only* these queries will be collected and 
inserted into the \code{status} table.  Additionally, these search queries
will be added to the \code{query_text} table if they do not already appear there. 
If \code{authentication.vector} is not supplied and no global 
\code{auth.vector} exists, this method returns an error.
}
\examples{

\dontrun{
auth.vector <- authorize_IT()

con <- twitter_database(
  "test2.sqlite",
  query.text.df = data.frame(
    query_text = c(
      "#throwbackthursday",
      "#worstfirstdate"
    )
  )
)
update_search(con)

DBI::dbDisonnect(con)
}
}
\seealso{
\code{\link{twitter_database}}, \code{\link{search_tweets_recursive}},
\code{\link{insert_statuses}}
}
