% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data.R
\name{update_users}
\alias{update_users}
\title{Update Users in Database}
\usage{
update_users(con, authentication.vector, user_id = NULL,
  screen_name = NULL, ...)
}
\arguments{
\item{con}{Twitter database connection (see \code{\link{twitter_database}}).}

\item{authentication.vector}{character vector with Twitter authentication
keys.  See \code{\link{authorize_IT}} or \code{\link{authorize_app}}.}

\item{user_id}{character vector of Twitter user_ids.}

\item{screen_name}{character vector of Twitter screen_names.  Ignored
if \code{user_id} is supplied.}

\item{...}{additional named parameters passed to \code{\link{user_lookup_recursive}}.}
}
\value{
\code{NULL} (Invisible).
}
\description{
Update User objects and user query table in a Twitter database
}
\details{
This function queries the Twitter API for user objects for users that
appear in the \code{query_users} table but have not been added to
the \code{user} table.  If a \code{screen_name} or \code{user_id} vector
is supplied, these users will be inserted into the \code{user} table 
and added into the \code{query_users} table, if they do not already appear.
If \code{authentication.vector} is not supplied and no global 
\code{auth.vector} exists, this method returns an error.
}
\examples{

\dontrun{
auth.vector <- authorize_IT()

con <- twitter_database("test1.sqlite")
update_users(con,screen_name=c("bostonglobe","nytimes"))

con2 <- twitter_database(
  "test2.sqlite",
  query.users.df = data.frame(
    screen_name = c(
      "bostonglobe",
      "nytimes"
    )
  )
)
update_users(con2)

DBI::dbDisonnect(con)
DBI::dbDisconnect(con2)
}
}
\seealso{
\code{\link{twitter_database}}, \code{\link{user_lookup_recursive}}, \code{\link{insert_users}}
}
