% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_model.R
\name{upload_query_text}
\alias{upload_query_text}
\title{Upload Search Text Query Table}
\usage{
upload_query_text(con, query.text.df, overwrite = FALSE)
}
\arguments{
\item{con}{a Twitter database connection.  See \code{twitter_database}.}

\item{query.text.df}{data frame.  See details and \code{\link{twitter_database}} documentation.}

\item{overwrite}{logical indicating whether to overwrite an existing \code{query_text} table in
the \code{conn} data connection.}
}
\value{
\code{NULL} (Invisible).
}
\description{
Upload a table of search text to a Twitter database connection
}
\details{
The \code{query.text.df} must be a data frame with a \code{query_text} column
containing character strings to execute in search queries.
Additional query category columns are optional.  A unique 'id' column will be added if
it is not included, as well as an additional mandatory {since_id} column.
See \code{\link{twitter_database}}.
}
\examples{
\dontrun{
users <- data.frame(
  screen_name=c(
    "nytimes",
    "latimes",
    "bostonglobe",
    "realDonaldTrump"
  ),
  category = c(
    "media",
    "media",
    "media",
    "politician"
  )
)
conn <- twitter_database("tweetanalysis.sqlite")
upload_query_users(conn,users)
}
}
\seealso{
\code{\link{twitter_database}}, \code{\link{update_search}}
}
