% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_model.R
\name{upload_query_users}
\alias{upload_query_users}
\title{Upload User Query Table}
\usage{
upload_query_users(con, query.users.df, overwrite = FALSE)
}
\arguments{
\item{con}{a Twitter database connection.  See \code{twitter_database}.}

\item{query.users.df}{data frame.  See details and \code{\link{twitter_database}} documentation.}

\item{overwrite}{logical indicating whether to overwrite an existing \code{query_user} table in
the \code{conn} data connection.}
}
\value{
\code{NULL} (Invisible).
}
\description{
Upload a table of users to a Twitter database connection
}
\details{
The \code{query.users.df} must be a data frame with either a \code{screen_name} column
or a \code{user_id} column containing Twitter screen names or user_ids for collection.
Additional user category columns are optional.  A unique 'id' column will be added if
it is not included.  Additional mandatory columns are \code{friends_collected}, 
\code{followers_collected}, and {since_id}.  These will all be automatically added and
populated if they are not included.
See \code{\link{twitter_database}}.
}
\examples{
\dontrun{
users <- data.frame(
  screen_name=c(
    "nytimes",
    "latimes",
    "bostonglobe",
    "realDonaldTrump"
  ),
  category = c(
    "media",
    "media",
    "media",
    "politician"
  )
)

conn <- twitter_database("tweetanalysis.sqlite")
upload_query_users(conn,users)
}
}
\seealso{
\code{\link{twitter_database}}, \code{\link{update_users}}, \code{\link{update_user_timelines}}
}
