% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{user_lookup}
\alias{user_lookup}
\title{Lookup Users}
\usage{
user_lookup(screen_name, user_id, authentication.vector,
  include_entities = TRUE, tweet_mode = "extended")
}
\arguments{
\item{screen_name}{character vector of up to 100 screen_names.  This is only used is \code{user_id} is missing.}

\item{user_id}{numeric or character vector of up to 100 user_ids.}

\item{authentication.vector}{character vector containing authentication tokens and secrets.
See \code{\link{authorize_app}} and \code{\link{authorize_IT}}.}

\item{include_entities}{logical indicating whether to include profile entities (e.g., urls)
as separate nodes in the returned json object.}

\item{tweet_mode}{character either 'extended' for full_text statuses or 'compat' for
140 character compatability.}
}
\value{
On success, an R list representation of the returned JSON object.  On failure, a warning
is thrown and the http response is returned, if any.
}
\description{
Get Twitter User Objects from the User Lookup API
}
\details{
This function calls \code{\link{twitter_request}} using the Twitter
users/lookup endpoint (see
\href{https://developer.twitter.com/en/docs/accounts-and-users/follow-search-get-users/api-reference/get-users-lookup}{User Lookup API Documentation}).
A user_id or screen_name vector is required.  If an authentication vector is not
provided and there is no globally defined \code{auth.vector}, an error is returned.
}
\examples{

\dontrun{
auth.vector <- authorize_IT()

user.list <- user_lookup(
  c(
  "realDonaldTrump",
  "SouthComWatch",
  "Stephenfaller"
  ),
  authentication.vector = auth.vector
)

cat(user.list[[1]]$description)
}
}
\seealso{
\code{\link{twitter_request}}, \code{\link{authorize_app}}, \code{\link{user_show}}
}
