% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_recursives.R
\name{user_lookup_recursive}
\alias{user_lookup_recursive}
\title{Recursive User Lookup Query}
\usage{
user_lookup_recursive(screen_name, user_id, data.connection,
  authentication.vector, include_entities = TRUE,
  tweet_mode = "extended", ...)
}
\arguments{
\item{screen_name}{character vector of screen_names.  This is only used is \code{user_id} is missing.}

\item{user_id}{numeric or character vector of user_ids.}

\item{data.connection}{a DBI connection to an RSQLite \code{\link{twitter_database}}.}

\item{authentication.vector}{character vector containing authentication tokens and secrets.
See \code{\link{authorize_app}} and \code{\link{authorize_IT}}.}

\item{include_entities}{logical indicating whether to include profile entities (e.g., urls)
as separate nodes in the returned json object.}

\item{tweet_mode}{character either 'extended' for full_text statuses or 'compat' for
140 character compatability.}

\item{...}{other named parameters passed to \code{\link{insert_users}}.}
}
\value{
\code{NULL} if \code{data.connection} is supplied, otherwise a list of user objects (lists).
}
\description{
Call the users/lookup API recursively
}
\details{
This function recursively calls the users/lookup endpoint (see
\href{https://developer.twitter.com/en/docs/accounts-and-users/follow-search-get-users/api-reference/get-users-lookup}{User Lookup API Documentation}).
This is useful for lists of more than 100 users.  For no more than 100 users, the
\code{\link{user_lookup}} function is preferred.  This method
includes a one second delay for each query in order to prevent exceeding the rate limit.

If the \code{authentication.vector} parameter is missing, this method checks the global
environment for \code{auth.vector}.  If this variable exists, the function will use
this variable as the \code{authentication.vector}.  If this variable does not exist, the
function will throw an error.
}
\examples{

## Not run: authenticate
# auth.vector <- authorize_IT()
user.ids.response <- get_followers_ids("realDonaldTrump")
user.ids <- user.ids.response$ids

users <- user_lookup_recursive(user_id=user.ids)

}
\seealso{
\code{\link{user_lookup}}, \code{\link{insert_users}}
}
