% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{wordcloud_plot}
\alias{wordcloud_plot}
\title{Word Cloud of Tweet Text}
\usage{
wordcloud_plot(con, start.date = NULL, end.date = NULL,
  where.criteria = NULL, caption = "", max.words = 50,
  stopword.langs = c("english", "spanish"),
  additional.stopwords = c("que", "las", "del", "con", "para", "una",
  "esta", "the", "https", "like", "can", "will", "also", "best", "must",
  "says", "amp"), file.name = NULL)
}
\arguments{
\item{con}{SQLite data connection to Twitter database (see \code{\link{twitter_database}}).}

\item{start.date}{character date (or datetime) of earliest statuses queried.}

\item{end.date}{character date.  Results will include statuses 
with creation times or dates strictly less than this parameter.}

\item{where.criteria}{character additional criteria to filter the status results, in 
SQLite syntax.}

\item{caption}{character caption to include on the plot.}

\item{max.words}{integer maximum number of words to plot.}

\item{stopword.langs}{character language librarys from the \code{tm} package to include.}

\item{additional.stopwords}{character additional stop words to remove from the text.}

\item{file.name}{character name of file to save the word cloud.  If \code{NULL}, the
word cloud is plotted on the screen.}
}
\value{
NULL (Invisible).
}
\description{
Create a word cloud of the combined text from tweets.
}
\details{
This function requires the word cloud and tm packages.  It generates a word cloud
from status text in the Twitter data connection.
}
\examples{

\dontrun{
auth.vector <- authorize_IT()

con <- twitter_database(
  "test.sqlite",
  query.users.df = data.frame(
    screen_name = c(
      "pb2pv",
      "zlisto"
    )
  ),
  query.text.df = data.frame(
    query_text = c(
      "#throwbackthursday",
      "#worstfirstdate"
    )
  )
)

update_user_timelines(con)
update_search(con)

wordcloud_plot(
  con,
  start.date = as.Date(Sys.time())-7,
  where.criteria = "user.screen_name = 'zlisto'",
  caption = "@zlisto"
)


DBI::dbDisonnect(con)
}
}
\seealso{
\code{\link{twitter_database}}, \code{\link{authorize_app}}
}
