% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter_authentication.R
\name{authorize_app}
\alias{authorize_app}
\title{Authorize Twitter Application}
\usage{
authorize_app(consumer.token, consumer.secret)
}
\arguments{
\item{consumer.token}{character Twitter application consumer token
(see \href{https://developer.twitter.com/en/docs/basics/authentication/overview/3-legged-oauth}{Twitter Oauth}).}

\item{consumer.secret}{character Twitter application consumer secret
(see \href{https://developer.twitter.com/en/docs/basics/authentication/overview/3-legged-oauth}{Twitter Oauth}).}
}
\value{
On successful authentication, a character vector containing the following named character strings:
\tabular{ll}{
\code{consumer.token} \tab The Twitter Application Consumer Token \cr
\code{consumer.secret} \tab The Twitter Application Consumer Secret \cr
\code{access.token} \tab The User's Access Token for the Twitter Application \cr
\code{access.secret} \tab The User's Access Secret for the Twitter Application.
}
On authentication failure, a warning is thrown and the failed http response object
is returned.
}
\description{
Authorize a Twitter application on a user account
}
\details{
The application consumer token and consumer secret must be provided.
These can be obtained by the application owner from the
\href{https://developer.twitter.com/en/apps}{Twitter Developer} site.
The user must log-in to Twitter (as directed) and authorize the
application.  The Twitter website will return a PIN, which the user
must enter in R to complete the authentication.
}
\section{Authentication Keys}{

The application consumer secret and user token secret are authentication passcodes
that should not be shared.  The authentication vector returned by this function
can be saved locally and reloaded for use at a later time, enabling the user
to skip the authentication step.
}

\section{Note}{

It is easy to hit <ENTER> in the console after calling this function before
entering your PIN, especially when copying and pasting code; this will
result in an authentication failure. To avoid this 
mistake, enter the \code{authorize_app} line manually into the console.  Then,
log on to Twitter and browse to the URL provided (if this does not happen automatically),
authorize the app in the browser, retrieve the PIN, and enter the PIN back into the
R Console.
}

\examples{
\dontrun{
## Consumer token and consumer secret are fake.
consumer.token <- "ArP09ecJ8ismgszHGuxis5p2tie6zLiLiWhgsWPbZxw3qpLSds"
consumer.secret <- "7N2e5hh1fYzFYfYib4OvRGv1zYAOpBGg2d2SK81qygnIYdtjnohehg2gx11HTInmCFF4uC9puLl"

# Be careful not to hit <ENTER> before entering your PIN following the next line.

auth.vector <- authorize_app(consumer.token,consumer.secret)

}
}
\seealso{
\code{\link{twitter_request}}, \code{\link{authorize_IT}}
}
