% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter_authentication.R
\name{authorize_IT}
\alias{authorize_IT}
\title{Authorize INFORMS-Tutorial Twitter Application}
\usage{
authorize_IT()
}
\value{
On successful authentication, a character vector containing the following named character strings:
\tabular{ll}{
\code{consumer.token} \tab The Twitter Application Consumer Token \cr
\code{consumer.secret} \tab The Twitter Application Consumer Secret \cr
\code{access.token} \tab The User's Access Token for the Twitter Application \cr
\code{access.secret} \tab The User's Access Secret for the Twitter Application.
}
On authentication failure, a warning is thrown and the failed http response object
is returned.
}
\description{
Authorize the INFORMS-Tutorial Twitter application for a user.
}
\details{
This is the easiest way to get started collecting and analyzing Twitter data.
All that is needed is a Twitter user account.
The application consumer token and consumer secret for the
INFORMS-Tutorial Application are provided in this package.
This function supplies these to the Twitter API to authorize this
*read only* application to query the Twitter API on the user's behalf.
See \href{http://zlisto.scripts.mit.edu/informs_tutorial/tutorial.html}{2015 INFORMS Twitter Mining Tutorial}
for more information about the origin of this Application.  This function
will require the user to log in to Twitter and authorize this application.
Once authorized, Twitter will provide the User with a PIN that must be entered
into R.  Using this PIN, R will complete the authentication process and return
the user's authentication credentials (access token and secret), which the user
can save locally for future use.
}
\section{Note}{

The application consumer secret and user token secret are authentication passcodes
that should not be shared.  In particular, the INFORMS-Tutorial consumer secret, while
provided as a part of this package, should not be shared or used outside of
the scope of mining Twitter data for the purposes of research and analyses that
conform to applicable laws and policies, including Twitter's terms of use.
}

\examples{
auth.vector <- authorize_IT()

}
\seealso{
\code{\link{twitter_request}}, \code{\link{authorize_app}}
}
