% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_recursives.R
\name{followers_ids_recursive}
\alias{followers_ids_recursive}
\title{Recursively Get Followers IDs}
\usage{
followers_ids_recursive(screen_name, user_id, data.connection,
  authentication.vector, stringify_ids = TRUE, tweet_mode = "extended",
  last.query.time = NULL, ...)
}
\arguments{
\item{screen_name}{character single screen_name.  This is only used is \code{user_id} is missing.}

\item{user_id}{numeric or character single user_id.}

\item{data.connection}{a DBI connection to an RSQLite \code{\link{twitter_database}}.}

\item{authentication.vector}{character vector containing authentication tokens and secrets.}

\item{stringify_ids}{logical indicating whether to return user_ids as strings.}

\item{tweet_mode}{character either 'extended' for full_text statuses or 'compat' for
140 character compatability.  This parameter has no effect if no \code{data.connection} is provided.}

\item{last.query.time}{POSIXct time of last query executed.  If not provided, a one minute delay will 
be added before the first query to prevent exceeding the Twitter rate limit.}

\item{...}{other named parameters passed to \code{\link{insert_users}}.}
}
\value{
character collected \code{user_id} if \code{data.connection} is supplied, otherwise a character or numeric vector of user_ids.
}
\description{
Recursively call the Followers IDs API to get all followers IDs
}
\details{
This function calls recursively calls the
followers/list endpoint (see
\href{https://developer.twitter.com/en/docs/accounts-and-users/follow-search-get-users/api-reference/get-followers-ids}{Followers IDs API Documentation}).
A user_id or screen_name is required.  If an authentication vector is not
provided and there is no globally defined \code{auth.vector}, an error is returned.  A *one minute* delay
is included for each query in order to avoid exceeding the rate limit.
If a \code{data.connection} is provided, user objects will also be collected and inserted.
}
\examples{

\dontrun{
auth.vector <- authorize_IT()

dma.followers <- followers_ids_recursive(
  "Nature_Island",
  authentication.vector = auth.vector
)

cat(
  dma.followers[1:20]
)
}
}
\seealso{
\code{\link{followers_ids}}, \code{\link{insert_followers}}, \code{\link{twitter_database}}
}
