% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{twitter_request}
\alias{twitter_request}
\title{Access Twitter REST Endpoint}
\usage{
twitter_request(auth.vector, endpoint.url, query.param.list = list())
}
\arguments{
\item{auth.vector}{character authentication vector (see Details).}

\item{endpoint.url}{character Twitter REST API endpoint URL.  See
\href{https://developer.twitter.com/en/docs/api-reference-index}{Twitter Documentation} for details, and
\href{https://developer.twitter.com/en/docs/tweets/search/api-reference/get-search-tweets.html}{Twitter Search API} for an example of where to find these URLs.}

\item{query.param.list}{named list giving the query parameters.  Names should
only correspond to parameters provided in the
\href{https://developer.twitter.com/en/docs/api-reference-index}{Twitter API Documentation}
specific to the endpoint being called.  Any parameter values with length greater than
one will be collapsed into a single, comma-separated string.}
}
\value{
httr response object.
}
\description{
Send an authenticated HTTPS request to a Twitter REST API endpoint.
}
\details{
This is the workhorse function of this package, as it is called by every 
function that queries the Twitter API.  It is useful for accessing API endpoints
that are not provided in this package.
The authentication vector must be provided.  It can be created using the
\code{\link{authorize_app}} or \code{\link{authorize_IT}} functions, or can be read from
a locally saved location if one has already been created.  This function
formulates an authenticated and signed HTTPS request and sends it to the
Twitter API endpoint specified.  Parameters vary among the endpoints; user
familiarity with the endpoint is assumed.
}
\examples{

\dontrun{
auth.vector <- load("auth-vector.RData")

resource.url <- "https://api.twitter.com/1.1/statuses/user_timeline.json"
params <- list(screen_name="realDonaldTrump",count="200",tweet_mode="extended")
httr.response <- twitter_request(auth.vector,resource.url,params)
}
}
\seealso{
\code{\link{authorize_app}},\code{\link{authorize_IT}}
}
